GraphQL server - audit the adding of projects giveth-2/issues#257
This is is mostly a code cleanup (#239 Excellence - as all changes were on the Dapp)

This uses the GATSBY_SITE_ID to determine the organisationId of the corresponding organisation in the database. This is the means where by many deployments of the dapp can use the same database, without seeing each other’s projects.

 https://github.com/Giveth/giveth-2/issues/257